Add useArticles hook to the Headlines Page
