+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import Headlines from "./Components/Pages/Headlines/index";
 
 import Header from "./Components/Header";
@@ -14,6 +14,10 @@ function App() {
       <BrowserRouter>
         <Header />
         <Routes>
+          <Route
+            path="/"
+            element={<Navigate replace to="/topheadlines/br" />}
+          />
           <Route path="/topheadlines/:country" element={<FeedArticle />} />
           <Route
             path="/topheadlines/:country/:category"
